Some refractors and added Mail and policies to Project

+++ app/Http/Controllers/ProjectsController.php
@@ -6,12 +6,19 @@ use Illuminate\Http\Request;
 use Illuminate\Filesystem\Filesystem;
 use App\Project;
 use App\Facebook;
+use App\Mail\ProjectCreated;
 
 class ProjectsController extends Controller
 {
+
+    public function __construct() {
+        $this->middleware('auth');
+    //    $this->middleware('can:update,project');
+    }
+
     public function index(){
 
-        $projects = Project::all();
+        $projects = Project::where('owner_id', auth()->id())->get();
         
         return view('projects.index', compact('projects'));
 
@@ -26,23 +33,38 @@ class ProjectsController extends Controller
 
     public function store(){
 
-        Project::create(
-            
-            request()->validate([
+        $validatedAttr = request()->validate(
+            [
                 'title' => ['required', 'min:5'],
 
                 'description' => ['required', 'min:10']
-            ])
+            ]
+        );
+
+        $validatedAttr['owner_id'] = auth()->id();
+
+
+        $project = Project::create($validatedAttr);
+
+        \Mail::to('david@localhost.com')->send(
+            new ProjectCreated($project)
         );
         
         return redirect('/projects');
     }
     
-    public function show(Project $project, Facebook $facebook)
+    public function show(Project $project)
     {
-
        /*  dd($facebook);
         dd(app('facebook')); */
+
+        /* if($project->owner_id != auth()->id()){
+            abort(403);
+        } */
+
+        // Using Polocy to check!
+
+       $this->authorize('update', $project);
 
         return view('projects.show', compact('project'));
         
@@ -50,18 +72,34 @@ class ProjectsController extends Controller
 
     public function edit(Project $project){
 
+       // abort_if($project->owner_id !== auth()->id(), 403);
+        /* if($project->owner_id != auth()->id()){
+            abort(403);
+        } */
+
+        //WIth the ProjectPolicy
+
+       $this->authorize('update', $project);
+
         return view('projects.edit', compact('project'));
         
     }
 
     public function update(Project $project){
 
-        $project->update(request(['title','description']));
+        $this->authorize('update', $project);
+        
+        $project->update(request()->validate([
+            'title' => ['required'],
+            'description' => ['required']
+            ])
+        );
 
         return redirect('/projects/'.$project->id);
     }
 
     public function destroy(Project $project){
+        $this->authorize('update', $project);
 
         $project->delete();
